Code Modification to Left Subtree Maximum Value
With learnings gathered throughout the part 2 video, I modified the code to configure them into setting the maximum value from the left subtree to replace the deleted root node with almost 2 childe nodes.

Also, I appended self.right to itself on code 109 for updating the binary tree repetitively when it is used as a print function on main method.